Move getCheckboxValues and removeFilter to search helper
