+++ app/controllers/search.js
@@ -21,21 +21,10 @@ const {
   getRegionLabel
 } = require('../helpers/gias')
 
-const getCheckboxValues = (name, data) => {
-  return name && (Array.isArray(name)
-    ? name
-    : [name].filter((name) => {
-        return name !== '_unchecked'
-      })) || data && (Array.isArray(data) ? data : [data])
-}
-
-const removeFilter = (value, data) => {
-  if (Array.isArray(data)) {
-    return data.filter(item => item !== value)
-  } else {
-    return null
-  }
-}
+const {
+  getCheckboxValues,
+  removeFilter
+} = require('../helpers/search')
 
 exports.search_get = async (req, res) => {
   delete req.session.data.search
