+++ app/controllers/support/placementSchool.js
@@ -1,16 +1,3 @@
-const Pagination = require('../../helpers/pagination')
-
-const {
-  getSchoolTypeOptions,
-  getSchoolTypeLabel,
-  getSchoolGroupOptions,
-  getSchoolGroupLabel,
-  getSchoolStatusOptions,
-  getSchoolStatusLabel,
-  getSchoolEducationPhaseOptions,
-  getSchoolEducationPhaseLabel
-} = require('../../helpers/gias')
-
 const {
   AcademicYear,
   PlacementSchool,
@@ -25,23 +12,25 @@ const {
   Sequelize
 } = require('../../models')
 
+const Pagination = require('../../helpers/pagination')
+
+const {
+  getSchoolTypeOptions,
+  getSchoolTypeLabel,
+  getSchoolGroupOptions,
+  getSchoolGroupLabel,
+  getSchoolStatusOptions,
+  getSchoolStatusLabel,
+  getSchoolEducationPhaseOptions,
+  getSchoolEducationPhaseLabel
+} = require('../../helpers/gias')
+
+const {
+  getCheckboxValues,
+  removeFilter
+} = require('../../helpers/search')
+
 const { Op } = require('sequelize')
-
-const getCheckboxValues = (name, data) => {
-  return name && (Array.isArray(name)
-    ? name
-    : [name].filter((name) => {
-        return name !== '_unchecked'
-      })) || data && (Array.isArray(data) ? data : [data])
-}
-
-const removeFilter = (value, data) => {
-  if (Array.isArray(data)) {
-    return data.filter(item => item !== value)
-  } else {
-    return null
-  }
-}
 
 const groupPlacementSchools = (rows) => {
   const grouped = {}
