+++ app/helpers/search.js
@@ -1,0 +1,20 @@
+const getCheckboxValues = (name, data) => {
+  return name && (Array.isArray(name)
+    ? name
+    : [name].filter((name) => {
+        return name !== '_unchecked'
+      })) || data && (Array.isArray(data) ? data : [data])
+}
+
+const removeFilter = (value, data) => {
+  if (Array.isArray(data)) {
+    return data.filter(item => item !== value)
+  } else {
+    return null
+  }
+}
+
+module.exports = {
+  getCheckboxValues,
+  removeFilter
+}
